+++ braindecode/datasets/bids/iterable.py
@@ -25,14 +25,13 @@ class BIDSIterableDataset(IterableDataset):
 
     Examples
     --------
-    >>> from braindecode.datasets import RecordDataset, BaseConcatDataset
-    >>> from braindecode.datasets.bids import BIDSIterableDataset, _description_from_bids_path
+    >>> from braindecode.datasets import BaseConcatDataset, RawDataset, RecordDataset
+    >>> from braindecode.datasets.bids import BIDSIterableDataset
     >>> from braindecode.preprocessing import create_fixed_length_windows
     >>>
     >>> def my_reader_fn(path):
     ...     raw = mne_bids.read_raw_bids(path)
-    ...     desc = _description_from_bids_path(path)
-    ...     ds = RawDataset(raw, description=desc)
+    ...     ds: RecordDataset = RawDataset(raw, description={"path": path.fpath})
     ...     windows_ds = create_fixed_length_windows(
     ...         BaseConcatDataset([ds]),
     ...         window_size_samples=400,
@@ -48,7 +47,8 @@ class BIDSIterableDataset(IterableDataset):
     Parameters
     ----------
     reader_fn : Callable[[mne_bids.BIDSPath], Sequence]
-        A function that takes a BIDSPath and returns a dataset.
+        A function that takes a BIDSPath and returns a dataset (e.g., a
+        RecordDataset or BaseConcatDataset of RecordDataset).
     pool_size : int
         The number of recordings to read and sample from.
     bids_paths : list[mne_bids.BIDSPath] | None